weight function in MovidaCore and NOGraph update

+++ movida/borghicremona/graph/Weight.java
@@ -3,5 +3,5 @@ package movida.borghicremona.graph;
 /* Instances of this type should be instantiated with a lambda expression in order to
    return the "weight" of an arch. */
 public interface Weight {
-	public int weight(Comparable nodeKeyA, Comparable nodeKeyB);
+	public Double weight(Comparable nodeKeyA, Comparable nodeKeyB);
 }
